feat: move util/phone-num-validation to root shared folder
