+++ src/app/utils/field-validation/validators/homeNoValidator.ts
@@ -1,14 +1,14 @@
 import { chain, left, right } from 'fp-ts/lib/Either'
 import { flow } from 'fp-ts/lib/function'
 
-import { ProcessedSingleAnswerResponse } from 'src/app/modules/submission/submission.types'
 import { IHomenoFieldSchema, OmitUnusedValidatorProps } from 'src/types/field'
 import { ResponseValidator } from 'src/types/field/utils/validation'
 
 import {
   isHomePhoneNumber,
   startsWithSgPrefix,
-} from '../../../../shared/util/phone-num-validation'
+} from '../../../../../shared/util/phone-num-validation'
+import { ProcessedSingleAnswerResponse } from '../../../modules/submission/submission.types'
 
 import { notEmptySingleAnswerResponse } from './common'
