+++ src/app/utils/field-validation/validators/mobileNoValidator.ts
@@ -1,14 +1,16 @@
 import { chain, left, right } from 'fp-ts/lib/Either'
 import { flow } from 'fp-ts/lib/function'
 
-import { ProcessedSingleAnswerResponse } from 'src/app/modules/submission/submission.types'
-import { IMobileFieldSchema, OmitUnusedValidatorProps } from 'src/types/field'
-import { ResponseValidator } from 'src/types/field/utils/validation'
-
 import {
   isMobilePhoneNumber,
   startsWithSgPrefix,
-} from '../../../../shared/util/phone-num-validation'
+} from '../../../../../shared/util/phone-num-validation'
+import {
+  IMobileFieldSchema,
+  OmitUnusedValidatorProps,
+} from '../../../../types/field'
+import { ResponseValidator } from '../../../../types/field/utils/validation'
+import { ProcessedSingleAnswerResponse } from '../../../modules/submission/submission.types'
 
 import { makeSignatureValidator, notEmptySingleAnswerResponse } from './common'
